libgstreamer-plugins-base1.0-dev: Add ${gir:Provides}
This adds the systematic names for the various included GIR XML files
(GstAllocator-1.0.gir ⇒ gir1.2-gstallocators-1.0-dev, and so on)
so that other packages can add Depends or Build-Depends on them
if required. It's also a step towards being able to add the nogir
build-profile in future, if desired.

Addresses Lintian tag: gir-package-name-does-not-match