Install GstAudio-1.0.gir into ${libdir}/gir-1.0
This allows libgstreamer-plugins-base1.0-dev to return to being
Multi-Arch: same without causing a regression similar to #1108412.

The cost of doing this is that this particular module will not be
available to static language bindings unless they search that path.
For example, cppgir (#1060906) and gir-to-d (#1060909) would need
workarounds.

According to codesearch there do not appear to be any bindings
in Debian that would be broken by this right now: in particular,
rust-gstreamer-audio-sys doesn't currently regenerate its bindings
from system GIR XML, and if it did, it could use a technique similar
to #1116500 to avoid needing to hard-code the path.

Closes: #1023591